gamebackend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
faster-whisper==1.1.1
rapidfuzz==3.9.0
python-multipart==0.0.9
python-dotenv==1.0.1
